Read mail files
No idea if this is idiomatic yet. Path was a chore to work with.
Rolling my own error enum to wrap things that could go wrong felt
reasonable, but a lot of boilerplate for this kind of thing. Maybe a
macro makes that feel better?

Also unclear if the owned types (PathBuf, String) are the right thing
for Document or if references make more sense? I tried that and it was
huge pain when it came to lifetimes. If moves are cheap maybe it doesn't
matter? Course, having to realloc contents seems like a bummer.